pyproject.toml
[project]
name = "research-os"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
